Move the non-primary CLI into utils.
